Refactoring utils files
