+++ aqme/utils.py
@@ -10,21 +10,14 @@ import time
 import getopt
 import numpy as np
 import glob
-import ast
 import yaml
-import pandas as pd
-import json
-import math
 from pathlib import Path
 from rdkit.Chem.rdMolAlign import GetBestRMS
 from rdkit.Chem.rdmolops import RemoveHs
-from rdkit import Geometry
-from rdkit.Chem import Mol, rdmolfiles
+from rdkit.Chem import Mol
 from rdkit.Chem import AllChem as Chem
 from aqme.argument_parser import set_options, var_dict
 from rdkit import RDLogger
-from aqme.xtb_to_json import read_json
-import rdkit
 
 GAS_CONSTANT = 8.3144621  # J / K / mol
 J_TO_AU = 4.184 * 627.509541 * 1000.0  # UNIT CONVERSION
@@ -35,222 +28,6 @@ time_run = time.strftime("%Y/%m/%d %H:%M:%S", time.localtime())
 aqme_ref = f"AQME v {aqme_version}, Alegre-Requena, J. V.; Sowndarya, S.; Perez-Soto, R.; Alturaifi, T. M.; Paton, R. S., 2022. https://github.com/jvalegre/aqme"
 
 RDLogger.DisableLog("rdApp.*")
-
-
-def get_boltz(energy):
-    """
-    Gets the boltz weights for a list of energies
-    """
-    energ = [number - min(energy) for number in energy]
-
-    boltz_sum = 0.0
-    for e in energ:
-        boltz_sum += math.exp(-e * J_TO_AU / GAS_CONSTANT / T)
-    weights = []
-
-    for e in energ:
-        weight = math.exp(-e * J_TO_AU / GAS_CONSTANT / T) / boltz_sum
-        weights.append(weight)
-
-    return weights
-
-
-def average_prop_mol(weights, prop):
-    """
-    Rtuerns the average properties for molecular properties
-    """
-    boltz_avg = 0.0
-    for i, p in enumerate(prop):
-        boltz_avg += p * weights[i]
-    return boltz_avg
-
-
-def average_prop_atom(weights, prop):
-    """
-    Rtuerns the average properties for molecular properties
-    """
-    boltz_avg = []
-    for i, p in enumerate(prop):
-        boltz_avg.append([number * weights[i] for number in p])
-    boltz_res = np.sum(boltz_avg, 0)
-    return boltz_res
-
-
-def get_boltz_avg_properties_xtb(
-    json_files,
-    name,
-    boltz_dir,
-    type,
-    nmr_atoms=None,
-    nmr_slope=None,
-    nmr_intercept=None,
-    nmr_experim=None,
-    mol=None,
-):
-    """
-    Gets the properties from json files and gives boltz averaged boltz properties
-    """
-    if type == "xtb":
-        mol_prop = [
-            "total energy",
-            "HOMO-LUMO gap/eV",
-            "electronic energy",
-            "Dipole module/D",
-            "Total charge",
-            "HOMO",
-            "LUMO",
-            "Fermi-level/eV",
-            "Total dispersion C6",
-            "Total dispersion C8",
-            "Total polarizability alpha",
-            "Total FOD",
-        ]
-        atom_prop = [
-            "dipole",
-            "partial charges",
-            "mulliken charges",
-            "cm5 charges",
-            "FUKUI+",
-            "FUKUI-",
-            "FUKUIrad",
-            "s proportion",
-            "p proportion",
-            "d proportion",
-            "Coordination numbers",
-            "Dispersion coefficient C6",
-            "Polarizability alpha",
-            "FOD",
-            "FOD s proportion",
-            "FOD p proportion",
-            "FOD d proportion",
-        ]
-    elif type == "nmr":
-        atom_prop = [
-            "NMR Chemical Shifts",
-        ]
-        if nmr_experim is not None:
-            exp_data = pd.read_csv(nmr_experim)
-
-    energy = []
-
-    for k, json_file in enumerate(json_files):
-        json_data = read_json(json_file)
-        if type == "xtb":
-            energy.append(json_data["total energy"])
-        elif type == "nmr":
-            energy.append(json_data["optimization"]["scf"]["scf energies"][-1])
-
-            json_data["properties"]["NMR"]["NMR Chemical Shifts"] = get_chemical_shifts(
-                json_data, nmr_atoms, nmr_slope, nmr_intercept
-            )
-            if nmr_experim is not None:
-                list_shift = json_data["properties"]["NMR"]["NMR Chemical Shifts"]
-                df = pd.DataFrame(
-                    list_shift.items(),
-                    columns=["atom_idx", "conf_{}".format(k + 1)],
-                )
-                df["atom_idx"] = df["atom_idx"] + 1
-                exp_data = exp_data.merge(df, on=["atom_idx"])
-        with open(json_file, "w") as outfile:
-            json.dump(json_data, outfile)
-
-    boltz = get_boltz(energy)
-
-    avg_json_data = {}
-    for prop in atom_prop:
-        prop_list = []
-        for json_file in json_files:
-            json_data = read_json(json_file)
-            if type == "xtb":
-                prop_list.append(json_data[prop])
-            if type == "nmr":
-                prop_list.append(json_data["properties"]["NMR"][prop].values())
-        avg_prop = average_prop_atom(boltz, prop_list)
-
-        if type == "nmr":
-            dictavgprop = {}
-            for i, key in enumerate(json_data["properties"]["NMR"][prop].keys()):
-                dictavgprop[key] = avg_prop[i]
-            avg_json_data[prop] = dictavgprop
-
-            if nmr_experim is not None:
-                list_shift = avg_json_data[prop]
-                df = pd.DataFrame(
-                    list_shift.items(),
-                    columns=["atom_idx", "boltz_avg"],
-                )
-                df["atom_idx"] = df["atom_idx"].astype(int) + 1
-                exp_data = exp_data.merge(df, on=["atom_idx"])
-                exp_data["error_boltz"] = abs(
-                    exp_data["experimental_ppm"] - exp_data["boltz_avg"]
-                )
-                exp_data.round(2).to_csv(
-                    nmr_experim.split(".csv")[0] + "_predicted.csv", index=False
-                )
-        elif type == "xtb":
-            avg_json_data[prop] = avg_prop.tolist()
-
-    if type == "xtb":
-        for prop in mol_prop:
-            prop_list = []
-            for json_file in json_files:
-                json_data = read_json(json_file)
-                prop_list.append(json_data[prop])
-            avg_prop = average_prop_mol(boltz, prop_list)
-            avg_json_data[prop] = avg_prop
-
-    final_boltz_file = str(boltz_dir) + "/" + name + "_boltz.json"
-    if mol is not None:
-        avg_json_data = get_rdkit_properties(avg_json_data, mol)
-    with open(final_boltz_file, "w") as outfile:
-        json.dump(avg_json_data, outfile)
-
-
-def get_rdkit_properties(avg_json_data, mol):
-    avg_json_data["NHOHCount"] = rdkit.Chem.Lipinski.NHOHCount(mol)
-    avg_json_data["FractionCSP3"] = rdkit.Chem.Lipinski.FractionCSP3(mol)
-    avg_json_data["NOCount"] = rdkit.Chem.Lipinski.NOCount(mol)
-    avg_json_data["NumAliphaticRings"] = rdkit.Chem.Lipinski.NumAliphaticRings(mol)
-    avg_json_data["NumAromaticRings"] = rdkit.Chem.Lipinski.NumAromaticRings(mol)
-    avg_json_data["NumHAcceptors"] = rdkit.Chem.Lipinski.NumHAcceptors(mol)
-    avg_json_data["NumHDonors"] = rdkit.Chem.Lipinski.NumHDonors(mol)
-    avg_json_data["NumHeteroatoms"] = rdkit.Chem.Lipinski.NumHeteroatoms(mol)
-    avg_json_data["NumRotatableBonds"] = rdkit.Chem.Lipinski.NumRotatableBonds(mol)
-
-    avg_json_data["TPSA"] = rdkit.Chem.Descriptors.TPSA(mol)
-    avg_json_data["MolLogP"] = rdkit.Chem.Descriptors.MolLogP(mol)
-    # avg_json_data["NumAmideBonds"] = rdkit.Chem.Descriptors.NumAmideBonds(mol)
-
-    return avg_json_data
-
-
-def get_chemical_shifts(json_data, nmr_atoms, nmr_slope, nmr_intercept):
-
-    if not isinstance(nmr_atoms, list):
-        nmr_atoms = ast.literal_eval(nmr_atoms)
-    if not isinstance(nmr_slope, list):
-        nmr_slope = ast.literal_eval(nmr_slope)
-    if not isinstance(nmr_intercept, list):
-        nmr_intercept = ast.literal_eval(nmr_intercept)
-
-    atoms = json_data["atoms"]["elements"]["number"]
-    tensor = json_data["properties"]["NMR"]["NMR isotopic tensors"]
-    shifts = {}
-    i = 0
-    for atom, ten in zip(atoms, tensor):
-        if atom in nmr_atoms:
-            # assigning values from arrays
-            index = nmr_atoms.index(atom)
-            slope_nuc = nmr_slope[index]
-            intercept_nuc = nmr_intercept[index]
-
-            scaled_nmr = (intercept_nuc - ten) / (-slope_nuc)
-            shifts[i] = scaled_nmr
-        else:
-            pass
-        i += 1
-
-    return shifts
 
 
 def run_command(command, outfile):
@@ -357,54 +134,6 @@ class Logger:
         self.log.close()
 
 
-def creation_of_dup_csv_cmin(cmin):
-
-    """
-    Generates a pandas.DataFrame object with the appropiate columns for the
-    conformational search and the minimization.
-
-    Parameters
-    ----------
-    cmin : str
-        Minimization method. Current valid methods are: ['xtb','ani']
-
-    Returns
-    -------
-    pandas.DataFrame
-    """
-
-    # Boolean aliases from args
-    is_xtb = cmin == "xtb"
-    is_ani = cmin == "ani"
-
-    # column blocks definitions
-
-    xtb_columns = [
-        "xTB-Initial-samples",
-        "xTB-energy-window",
-        "xTB-initial_energy_threshold",
-        "xTB-RMSD-and-energy-duplicates",
-        "xTB-Unique-conformers",
-    ]
-    ANI_columns = [
-        "ANI-Initial-samples",
-        "ANI-energy-window",
-        "ANI-initial_energy_threshold",
-        "ANI-RMSD-and-energy-duplicates",
-        "ANI-Unique-conformers",
-    ]
-    end_columns = ["CMIN time (seconds)", "Overall charge"]
-
-    # Check Minimization Method
-    if is_ani:
-        columns = ANI_columns
-    if is_xtb:  # is_ani and is_xtb will not happen, but this is what was written
-        columns = xtb_columns
-
-    columns += end_columns
-    return pd.DataFrame(columns=columns)
-
-
 def move_file(destination, source, file):
     """
     Moves files from the source folder to the destination folder and creates
@@ -497,197 +226,6 @@ def get_info_input(file):
     return atoms_and_coords, charge, mult
 
 
-def smi_to_mol(
-    smi,
-    name,
-    program,
-    log,
-    constraints_atoms,
-    constraints_dist,
-    constraints_angle,
-    constraints_dihedral,
-):
-
-    complex_ts = False
-    smi = smi.split(".")
-    if (
-        len(smi) > 1
-        or len(constraints_atoms) != 0
-        or len(constraints_dist) != 0
-        or len(constraints_angle) != 0
-        or len(constraints_dihedral) != 0
-    ):
-        if program not in ["crest"]:
-            log.write(
-                "\nx  Program not supported for conformer generation of complexes and TSs! Specify: program='crest' for complexes"
-            )
-            sys.exit()
-
-        (
-            mol,
-            constraints_atoms,
-            constraints_dist,
-            constraints_angle,
-            constraints_dihedral,
-        ) = nci_ts_mol(
-            smi,
-            name,
-            constraints_atoms,
-            constraints_dist,
-            constraints_angle,
-            constraints_dihedral,
-        )
-        complex_ts = True
-
-    else:
-        params = Chem.SmilesParserParams()
-        params.removeHs = False
-        try:
-            mol = Chem.MolFromSmiles(smi[0], params)
-        except Chem.AtomValenceException:
-            log.write(
-                f"\nx  The SMILES string provided ( {smi[0]} ) contains errors. For example, N atoms from ligands of metal complexes should be N+ since they're drawn with four bonds in ChemDraw, same for O atoms in carbonyl ligands, etc.\n"
-            )
-            sys.exit()
-
-    return (
-        mol,
-        constraints_atoms,
-        constraints_dist,
-        constraints_angle,
-        constraints_dihedral,
-        complex_ts
-    )
-
-
-def rdkit_sdf_read(file, args):
-    """
-    Reads sdf files and stops the execution if the file was not accesible.                                                                                                                                                                                      rdkit.Chem.Mol objects
-    """
-    
-    inmols = Chem.SDMolSupplier(file, removeHs=False)
-
-    if inmols is None:
-        args.log.write(f"Could not open {file}")
-        args.log.finalize()
-        sys.exit()
-    return inmols
-
-
-def nci_ts_mol(
-    smi,
-    name,
-    constraints_atoms,
-    constraints_dist,
-    constraints_angle,
-    constraints_dihedral,
-):
-    if constraints_atoms is not None:
-        constraints_atoms = [[float(y) for y in x] for x in constraints_atoms]
-    if constraints_dist is not None:
-        constraints_dist = [[float(y) for y in x] for x in constraints_dist]
-        constraints_dist = np.array(constraints_dist)
-    if constraints_angle is not None:
-        constraints_angle = [[float(y) for y in x] for x in constraints_angle]
-        constraints_angle = np.array(constraints_angle)
-    if constraints_dihedral is not None:
-        constraints_dihedral = [[float(y) for y in x] for x in constraints_dihedral]
-        constraints_dihedral = np.array(constraints_dihedral)
-
-    molsH = []
-    mols = []
-    for m in smi:
-        mols.append(Chem.MolFromSmiles(m))
-        molsH.append(Chem.AddHs(Chem.MolFromSmiles(m)))
-
-    for m in molsH:
-        Chem.EmbedMultipleConfs(m, numConfs=1)
-    for m in mols:
-        Chem.EmbedMultipleConfs(m, numConfs=1)
-
-    coord = [0.0, 0.0, 5.0]
-    molH = molsH[0]
-    for _, fragment in enumerate(molsH[1:]):
-        offset_3d = Geometry.Point3D(coord[0], coord[1], coord[2])
-        molH = Chem.CombineMols(molH, fragment, offset_3d)
-        coord[1] += 5
-        Chem.SanitizeMol(molH)
-
-    coord = [0.0, 0.0, 5.0]
-    mol = mols[0]
-    for _, fragment in enumerate(mols[1:]):
-        offset_3d = Geometry.Point3D(coord[0], coord[1], coord[2])
-        mol = Chem.CombineMols(mol, fragment, offset_3d)
-        coord[1] += 5
-        Chem.SanitizeMol(mol)
-    mol = Chem.AddHs(mol)
-
-    mol = Chem.ConstrainedEmbed(mol, molH)
-
-    atom_map = []
-    for atom in mol.GetAtoms():
-        atom_map.append(atom.GetAtomMapNum())
-
-    max_map = max(atom_map)
-    for a in mol.GetAtoms():
-        if a.GetSymbol() == "H":
-            max_map += 1
-            a.SetAtomMapNum(int(max_map))
-
-    nconstraints_atoms = []
-    if constraints_atoms is not None:
-        for _, ele in enumerate(constraints_atoms):
-            for atom in mol.GetAtoms():
-                if ele == atom.GetAtomMapNum():
-                    nconstraints_atoms.append(float(atom.GetIdx()) + 1)
-        nconstraints_atoms = np.array(nconstraints_atoms)
-
-    nconstraints_dist = []
-    if constraints_dist is not None:
-        for _, r in enumerate(constraints_dist):
-            nr = []
-            for _, ele in enumerate(r[:2]):
-                for atom in mol.GetAtoms():
-                    if ele == atom.GetAtomMapNum():
-                        nr.append(float(atom.GetIdx()) + 1)
-            nr.append(r[-1])
-            nconstraints_dist.append(nr)
-        nconstraints_dist = np.array(nconstraints_dist)
-
-    nconstraints_angle = []
-    if constraints_angle is not None:
-
-        for _, r in enumerate(constraints_angle):
-            nr = []
-            for _, ele in enumerate(r[:3]):
-                for atom in mol.GetAtoms():
-                    if ele == atom.GetAtomMapNum():
-                        nr.append(float(atom.GetIdx()) + 1)
-            nr.append(r[-1])
-            nconstraints_angle.append(nr)
-        nconstraints_angle = np.array(nconstraints_angle)
-
-    nconstraints_dihedral = []
-    if constraints_dihedral is not None:
-        for _, r in enumerate(constraints_dihedral):
-            nr = []
-            for _, ele in enumerate(r[:4]):
-                for atom in mol.GetAtoms():
-                    if ele == atom.GetAtomMapNum():
-                        nr.append(float(atom.GetIdx()) + 1)
-            nr.append(r[-1])
-            nconstraints_dihedral.append(nr)
-        nconstraints_dihedral = np.array(nconstraints_dihedral)
-
-    return (
-        mol,
-        nconstraints_atoms,
-        nconstraints_dist,
-        nconstraints_angle,
-        nconstraints_dihedral,
-    )
-
-
 def rules_get_charge(mol, args, type):
     """
     Automatically sets the charge for metal complexes
@@ -776,6 +314,7 @@ def rules_get_charge(mol, args, type):
             charge = Chem.GetFormalCharge(mol)
             return charge, metal_found
 
+
 def substituted_mol(self, mol, checkI):
     """
     Returns a molecule object in which all metal atoms specified in args.metal_atoms
@@ -804,36 +343,6 @@ def substituted_mol(self, mol, checkI):
                     atom.SetFormalCharge(formal_charge)
 
     return self.args.metal_idx, self.args.complex_coord, self.args.metal_sym
-
-
-def getDihedralMatches(mol, heavy):
-    # this is rdkit's "strict" pattern
-    pattern = r"*~[!$(*#*)&!D1&!$(C(F)(F)F)&!$(C(Cl)(Cl)Cl)&!$(C(Br)(Br)Br)&!$(C([CH3])([CH3])[CH3])&!$([CD3](=[N,O,S])-!@[#7,O,S!D1])&!$([#7,O,S!D1]-!@[CD3]=[N,O,S])&!$([CD3](=[N+])-!@[#7!D1])&!$([#7!D1]-!@[CD3]=[N+])]-!@[!$(*#*)&!D1&!$(C(F)(F)F)&!$(C(Cl)(Cl)Cl)&!$(C(Br)(Br)Br)&!$(C([CH3])([CH3])[CH3])]~*"
-    qmol = Chem.MolFromSmarts(pattern)
-    matches = mol.GetSubstructMatches(qmol)
-
-    # these are all sets of 4 atoms, uniquify by middle two
-    uniqmatches = []
-    seen = set()
-    for (a, b, c, d) in matches:
-        if (b, c) not in seen and (c, b) not in seen:
-            if heavy:
-                if (
-                    mol.GetAtomWithIdx(a).GetSymbol() != "H"
-                    and mol.GetAtomWithIdx(d).GetSymbol() != "H"
-                ):
-                    seen.add((b, c))
-                    uniqmatches.append((a, b, c, d))
-            if not heavy:
-                if (
-                    mol.GetAtomWithIdx(c).GetSymbol() == "C"
-                    and mol.GetAtomWithIdx(d).GetSymbol() == "H"
-                ):
-                    pass
-                else:
-                    seen.add((b, c))
-                    uniqmatches.append((a, b, c, d))
-    return uniqmatches
 
 
 def set_metal_atomic_number(mol, metal_idx, metal_sym):
@@ -1140,6 +649,7 @@ def cclib_atoms_coords(cclib_data):
     Function to convert atomic numbers and coordinate arrays from cclib into
     a format compatible with QPREP.
     """
+
     atom_numbers = cclib_data["atoms"]["elements"]["number"]
     atom_types = []
     per_tab = periodic_table()
